app/requirements.txt
flask
flask-migrate
psycopg2-binary
flask-sqlalchemy==2.5.1
paho-mqtt

